requirements.txt
matplotlib==3.7.1
numpy==1.25.2
pytorch==2.2.1
pyyaml==6.0.1
scikit-learn==1.2.2
